# Add aggregate and transfer performance tests