--- a/spec/double_entry/performance/reporting/aggregate_performance_spec.rb
+++ b/spec/double_entry/performance/reporting/aggregate_performance_spec.rb
@@ -1,21 +1,19 @@
 module DoubleEntry
   module Reporting
-
     RSpec.describe Aggregate do
       include PerformanceHelper
       let(:user) { User.make! }
 
       context '23,520 transfers spread out over 8 years' do
-        before do  #setup takes about 191s
+        before do # setup takes about 191s
           (2008..2015).each do |year|
             (1..12).each do |month|
               (1..28).each do |day|
                 Timecop.freeze Time.local(year, month, day) do
                   10.times { perform_deposit user, 1_00 }
-                  #TODO perform_deposit_with_metadata
+                  # TODO: perform_deposit_with_metadata
                 end
               end
-
             end
           end
         end
@@ -33,20 +31,19 @@ module DoubleEntry
           result = stop_profiling('aggregate')
           expect(total_time(result)).to be_faster_than(:local => 2.5, :ci => 2.5)
         end
-
       end
 
       context '23,520 transfers in a single day' do
-        before do #setup takes about 170s
+        before do # setup takes about 170s
           Timecop.freeze Time.local(2015, 06, 30) do
-            23520.times { perform_deposit user, 1_00 }
-            #TODO perform_deposit_with_metadata
+            23_520.times { perform_deposit user, 1_00 }
+            # TODO: perform_deposit_with_metadata
           end
         end
 
         it 'should calculate monthly all_time ranges quickly' do
           start_profiling
-          #TODO aggregate with metadata filter
+          # TODO: aggregate with metadata filter
           Reporting.aggregate(
             :sum, :savings, :bonus,
             :range => TimeRange.make(:year => 2015, :month => 06, :range_type => :all_time)
@@ -54,7 +51,6 @@ module DoubleEntry
           result = stop_profiling('aggregate')
           expect(total_time(result)).to be_faster_than(:local => 0.1, :ci => 0.1)
         end
-
       end
     end
   end