Performance tests for creating and aggregating transfers. These need to be augmented with meta-data once the meta-data work is underway.