@macosgrove 2015-07-29
Just like the reports_controller_performance_spec, the number of transfers makes no difference to the speed of the aggregation:
```
20,000 => 524ms
10,000 => 573ms
1,000   => 486ms
100     => 608ms
10       => 509ms
1         => 473ms
```