@rabidcarrot 2015-07-30
Looks good to me @macosgrove  +1
I'm sure we'll tweak these as we go